Fetch course data from api
