+++ src/stylesheets/index.css
@@ -1,3 +1,3 @@
-/* @tailwind base;
+@tailwind base;
 @tailwind components;
-@tailwind utilities; */
+@tailwind utilities;
